+++ plugins/lighthouse/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @backstage/plugin-lighthouse
 
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [b6557c098]
+- Updated dependencies [2527628e1]
+- Updated dependencies [6011b7d3e]
+- Updated dependencies [e3bd9fc2f]
+- Updated dependencies [1c69d4716]
+- Updated dependencies [d8d5a17da]
+- Updated dependencies [83b6e0c1f]
+- Updated dependencies [1665ae8bb]
+- Updated dependencies [04f26f88d]
+- Updated dependencies [ff243ce96]
+- Updated dependencies [e3bd9fc2f]
+  - @backstage/core-api@0.2.5
+  - @backstage/core@0.4.0
+  - @backstage/plugin-catalog@0.2.6
+  - @backstage/config@0.1.2
+  - @backstage/catalog-model@0.5.0
+  - @backstage/theme@0.2.2
+
 ## 0.2.4
 
 ### Patch Changes